+++ protocol/trojan/outbound.go
@@ -29,14 +29,14 @@ func (o *TrojanOutboundConnSession) SetMeter(meter stat.TrafficMeter) {
 
 func (o *TrojanOutboundConnSession) Write(p []byte) (int, error) {
 	n, err := o.rwc.Write(p)
-	o.meter.Count(uint64(n), 0)
+	o.meter.Count(n, 0)
 	o.sent += uint64(n)
 	return n, err
 }
 
 func (o *TrojanOutboundConnSession) Read(p []byte) (int, error) {
 	n, err := o.rwc.Read(p)
-	o.meter.Count(0, uint64(n))
+	o.meter.Count(0, n)
 	o.recv += uint64(n)
 	return n, err
 }
